refactor: colocate audio source type with source mappings
