+++ src/lib/recorder/audio-sources.ts
@@ -1,6 +1,14 @@
-import type { AudioPlaybackSource } from "./audio-buffer-playback.ts";
 import type { AudioTrackState } from "./runtime.ts";
 import type { TakeRegion } from "./take.ts";
+
+/** A buffer slice placed on the timeline, with all times in seconds. */
+export interface AudioPlaybackSource {
+  buffer: AudioBuffer;
+  /** Timeline position corresponding to buffer time zero. */
+  timelineOffset: number;
+  timelineStart: number;
+  timelineEnd: number;
+}
 
 export function getAudioTrackSources(
   track: AudioTrackState,
